Model repackaging Moved `Statistics` to `statistics/domain/model/` package. Moved `Task` to `tasks/domain/model/` package.
