+++ todoapp/app/src/main/java/com/example/android/architecture/blueprints/todoapp/statistics/domain/usecase/GetStatistics.java
@@ -3,10 +3,10 @@ package com.example.android.architecture.blueprints.todoapp.statistics.domain.us
 import android.support.annotation.NonNull;
 
 import com.example.android.architecture.blueprints.todoapp.UseCase;
-import com.example.android.architecture.blueprints.todoapp.data.Task;
+import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Task;
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource;
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository;
-import com.example.android.architecture.blueprints.todoapp.statistics.domain.Statistics;
+import com.example.android.architecture.blueprints.todoapp.statistics.domain.model.Statistics;
 
 import java.util.List;
