+++ todoapp/app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/TasksContract.java
@@ -19,7 +19,7 @@ package com.example.android.architecture.blueprints.todoapp.tasks;
 import android.support.annotation.NonNull;
 
 import com.example.android.architecture.blueprints.todoapp.BaseView;
-import com.example.android.architecture.blueprints.todoapp.data.Task;
+import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Task;
 import com.example.android.architecture.blueprints.todoapp.BasePresenter;
 
 import java.util.List;
