+++ todoapp/app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/domain/filter/TaskFilter.java
@@ -16,7 +16,7 @@
 
 package com.example.android.architecture.blueprints.todoapp.tasks.domain.filter;
 
-import com.example.android.architecture.blueprints.todoapp.data.Task;
+import com.example.android.architecture.blueprints.todoapp.tasks.domain.model.Task;
 
 import java.util.List;
