+++ todoapp/app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/domain/model/Task.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.android.architecture.blueprints.todoapp.data;
+package com.example.android.architecture.blueprints.todoapp.tasks.domain.model;
 
 import android.support.annotation.Nullable;
